+++ libs/langchain-aws/src/types.ts
@@ -1,0 +1,16 @@
+import type { ToolChoice } from "@aws-sdk/client-bedrock-runtime";
+import type { AwsCredentialIdentity, Provider } from "@aws-sdk/types";
+import { ConverseCommand } from "@aws-sdk/client-bedrock-runtime";
+
+export type CredentialType =
+  | AwsCredentialIdentity
+  | Provider<AwsCredentialIdentity>;
+
+export type ConverseCommandParams = ConstructorParameters<
+  typeof ConverseCommand
+>[0];
+
+export type BedrockToolChoice =
+  | ToolChoice.AnyMember
+  | ToolChoice.AutoMember
+  | ToolChoice.ToolMember;
